# metal: optimize matrix multiplication kernel