- Fixed field access in ggml_metal_kargs_mul_mm struct to use correct dimensions
- Improved shared memory access patterns in tiled matrix multiplication
- Added proper bounds checking for edge cases
- Enhanced thread synchronization for better performance

Performance improvements on M1 Max:
- Prompt processing (pp512): 437.30 → 5426.66 tokens/s (1140% increase!)
- Token generation (tg128): 58.58 → 56.56 tokens/s (stable)

Build: eb394993 (5549)

*Make sure to read the [contributing guidelines](https://github.com/ggml-org/llama.cpp/blob/master/CONTRIBUTING.md) before submitting a PR*